requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pydantic==1.10.8
PyYAML==6.0
requests==2.31.0
superagi-tools==1.0.6
typing_extensions==4.7.1
urllib3==2.0.4
